pyproject.toml
[project]
name = "network-speed-recorder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psutil>=7.0.0",
    "speedtest-cli>=2.1.3",
]
